requirements.txt
Django==4.1.7
djangorestframework-simple-apikey==0.1.1